TASK Use make for developer setup and asdf for nodejs version management
The NHS repository template contains makefiles for managing local
developer setup and build tasks, and uses a tool called asdf for
version management. Update readme with prerequisites to get make and
asdf working.

asdf does not support the latest version of yarn so preserved volta for
the time being until the decision of whether to continue with yarn is
completed.