VIA-2 MD/DB Restore mocking of content layer in vaccine page unit tests
Re-introducing the concept of mocking allows testing of when errors are
thrown in the content stack and de-couples the page unit tests from
depending on the content stack being implemented correctly. A separate
file, integration.test exists for testing that the real content stack
and the individual vaccine pages work together correctly.