VIA-2 Add optional OverviewInsetText to each Vaccine
Uses the VaccineInfo object to add some custom properties which render
for some vaccines and not others.

For this story this piece of insetText is hardcoded because it cannot
be read from the content API; the content response does not contain a
HealthAspect or anything similar on this element so we cannont find /
locate it in the content data.

Ideally the fields in the VaccineInfo object could be booleans telling
the ContentStylingService what to extract from the ContentAPI JSON.