VIA-182 AJ/DB: Modify hyperlinks from ContentAPI copy to open in new tab
DomPurifier set to add target=_blank and rel='noopener' to all
hyperlinks. Code refactored to only run sanitizeHtml on content from the
Content APi (in the styling service) and not in the componenets
directly, assuming all content rendered in the components originated
from our codebase and would therefore already be sanitised in earlier
steps.
Reference: https://owasp.org/www-community/attacks/Reverse_Tabnabbing
https://github.com/cure53/DOMPurify/issues/317
OWASP notes the noopener property may not be nessasary in modern
browsers as of 2023, assumption made that older browsers will need to
be supported.