VIA-264 AJ/DB Extract RootID from amznTraceId, refactor shared code in logger
Amazon documentation describes the amznTraceId field as chaning several
properties together, including the traceId of the current place in the
execution. Extract this rootId as it is more readable than the whole
string.

Also extracts logic common to the loggers for both the NextJS Middleware
Edge Runtime and Node into a single shared location.