VIA-331 Fix Wiremock EliD stub to return HTTP error codes when expected
By default if a URL matches multiple Wiremock rules it will use the
latest one declared in the file. Use the 'priority' field to override
this behaviour; default priority is 5, highest priority is 1, so setting
the EliD fallthrough case to a big number should ensure the fallthrough
is always a last resort.