VIA-378 Make CacheHydrator Github action fail if error thrown during lambda execution
AWS CLI treats any lambda invocation as a success if the lambda actually
ran, regardless of any business logic errors thrown during the lambda's
execution (see https://github.com/aws/aws-cli/issues/3470 ).

Inspect the contents of the JSON response to determine if an error
occurred and fail the build if so.