VIA-357 MD/DB Fix middleware adding extra request headers to only modify request, not response
In the previous implementation the headers sent into the
NextResponse.next method get applied to both the inbound reuqest AND
override the response headers as well. This unintended manipulation of
the response headers causes errors when calling Server Actions, where
next.js expects a particular content-type in the response and the
framework will throw an unhandled error if the header is not set as
expected.