VIA-357 MD/DB Add event listener for unhandled asynchronous client side error logging
Note that this will result in all unhandled errors in the client
outputting a log line stating that an unhandled error has happened.
COntext on what happened is impossible to pass in without accepting raw
text from an endpoint which may present a security risk. Error types can
be intercepted and directly specified in client-error-logger if
required.