VIA-411 Prevent browser caching NHS App redirect response for unauthenticated users
Fixes behaviour seen in Chrome browser where if a logged out user
has attempted to access the hub page and been automatically redirected
to NHS App (with a 307 response), all future attempts to access the hub
page from a route that triggers a http call (e.g. pressing the back
button from /cookie-policy or /accessibility) will follow the same
redirect regardless of the user's logged-in status. This results in the
user sometimes ending up at NHS App when we still expected them to be in
VitA.

Caused by the browser caching the 307 response and serving it for
subsequent http calls to that URL. Nothing is shown in the
application logs because this behaviour happens in the browser and
VitA's application logic is never invoked.

Navigating back from Vaccine pages was not affected as these pages
behave like a single-page-webapp and there is no http call required to
get back to the /check-and-book-rsv route from here; the navigation
happens in-memory in the javascript bundle.