VIA-598 Prevent E2E tests from being re-run if another version has been deployed in the meantime
Due to how concurrency rules work in Github it is entirely possible that
the deploy and e2e steps from multiple workflows will run in a arbitrary
order. Additionally, devs require the ability to re-run a failed E2E run
but there is no guarantee that another workflow hasn't deployed a
different version of the app since, resulting in the re-run being
executed against the wrong build and an erroneous result.

Start the E2E test job with a check that the build it expects to be
deployed actually is, and fail the build if not.