VIA-608 Render care card headings at level specified in Content-API
Fixes issue on the Rotavirus page where the care card was rendering
without a heading. In the content-api response this care card heading is
at level h4 because it sits beneath another subheading, which we have
not seen on other pages up until this point.

Updates the content styling service to preserve the heading level and
use it when rendering the react component.

Heading level is optional to preserve compatibility with other
non-content api usages of the care card (EliD, rsv-pregnancy, etc); if
heading level is not specified the react component library will render
at the default level as it did before (h2 at time of writing).

Note that incorrect heading levels would not have been obviously visible
in the UI as a single css style was applied which sets the text-size.
However, a screenreader would have noticed that the h2/h3/h4 elements
were appearing in the wrong structure.