VIA-630 Move age calculation from getSession to updateToken on login
Placing the age calculation on getSession had the unintended side
effect of being re-evaluated every time NextAuth loaded the session
(which it does often; twice every minute), and any errors with
the dateOfBirth resulted in a single user outputing many errors in the
logs.

Instead move the age calculation to the point when a user logs in and
user details are being persisted to the token; store age group on the
token for use by the session.