VIA-630 MD/DB Redirect user to service failure page if age unknown on hub page
Restores changes from 70350c7, moved up a layer to the hub page itself
so that the age based card component will still continue to load for the
other age ranges that have not been implemented yet (returns empty
element but still allows the hub to be loaded).

It was considered whether the error should be thrown in the session loader
instead and let the default existing error handling run; however this
redirects the user back to NHS app, which was not the required behaviour
for the age based hub feature.